% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCoxSplinePlot.R
\name{doCoxSplinePlot}
\alias{doCoxSplinePlot}
\title{Smoothed spline plotting from a CoxPH model}
\usage{
doCoxSplinePlot(
  x = stop("Need to provide vector of x data: x=..."),
  fit = stop("Need to provide CoxPH model with pspline term: fit=..."),
  x.lab = "",
  title = "",
  subtitle = "",
  returnValues = FALSE,
  legendPos = NULL,
  distBoxPlot = FALSE,
  yMax = NULL,
  xLims = c(NULL, NULL),
  bg_box_col = "grey98",
  null_line_type = 2,
  null_line_col = "grey60",
  null_line_wd = 1.5,
  h_lines_pos = NULL,
  h_lines_type = 2,
  h_lines_col = "grey60",
  h_lines_wd = 1,
  v_lines_pos = NULL,
  v_lines_type = 2,
  v_lines_col = "grey60",
  v_lines_wd = 1,
  ...
)
}
\arguments{
\item{x}{{required} vector of x-values (the primary exposure in the Cox model -- used with pspline() function)}

\item{fit}{{required} CoxPH object -- fitted model}

\item{x.lab}{label for x-axis, e.g. "Serum albumin (g/dL)" - whatever your exposure is}

\item{title}{title for the plot}

\item{subtitle}{additional subtitle}

\item{returnValues}{return a data.frame of the x and y values (default is FALSE)}

\item{legendPos}{default is NO results legend. Specify "topleft" or similar to show the results on the plot}

\item{distBoxPlot}{show bowplot of distribution? Default is FALSE}

\item{yMax}{manually set the y-axis maximum? (minimum always 0)}

\item{xLims}{manually set the x-axis limits? (needs to be a vector of length 2 - min or max can be NULL)}

\item{bg_box_col}{specific background colour of plot (default = grey98)}

\item{null_line_type}{horizontal reference line type (default = 2)}

\item{null_line_col}{horizontal reference line colour (default = grey60)}

\item{null_line_wd}{horizontal reference line width (default = 1.5)}

\item{h_lines_pos}{vector of y axis positions for grid lines if desired (default = NULL)}

\item{h_lines_type}{type of lines for y axis (default = 2)}

\item{h_lines_col}{colour of lines for y axis (default = grey60)}

\item{h_lines_wd}{width of lines for y axis (default = 1)}

\item{v_lines_pos}{vector of x axis positions for grid lines if desired (default = NULL)}

\item{v_lines_type}{type of lines for x axis (default = 2)}

\item{v_lines_col}{colour of lines for x axis (default = grey60)}

\item{v_lines_wd}{width of lines for x axis (default = 1)}

\item{...}{other plot options (e.g. cex)}
}
\value{
A plot of the spline term in a CoxPH model
}
\description{
Plots a smoothed spline curve for the hazard ratio (y) for each value of the primary exposure (x) from a Cox proportional hazard model.

 Function adapted from analyses by Ambarish Dutta:
         Dutta et al, 2013. Uric acid measurement improves prediction of cardiovascular mortality in later life. 
         J Am Geriatr Soc. 2013 Mar;61(3):319-26.
         http://www.ncbi.nlm.nih.gov/pubmed/23496291

v0.190611
}
\examples{
# create the 'survival object'
surv.death <- na.omit( Surv(data$age_death , data$dead) )
      
# do Cox model -- including covariates -- primary independent variable (exposure of interest) must be pspline()
pham.fit <- coxph( surv.death ~ pspline(data$albumin, df=4) + data$sex )
      
# use doCoxSplinePlot() to plot the smoothed curve (including CI's) for the Cox model
doCoxSplinePlot(x=data$albumin, 
                fit=pham.fit, 
                x.lab="Albumin (g/dL)", 
                title="Circulating albumin and risk of mortality")

}
\author{
Luke Pilling
}
