% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{extreme_p}
\alias{extreme_p}
\alias{get_extreme_p}
\title{Get extreme p-value}
\usage{
get_extreme_p(z)
}
\arguments{
\item{z}{a z (or t) statistic}
}
\value{
Returns a string (p-value)
}
\description{
Return p-value for any z (or t) statistic as a string
}
\examples{
z = 50
get_extreme_p(z)
#>  [1] "2.16e-545"

}
\author{
The internet (anon)
}
