% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_neglog10_p}
\alias{get_neglog10_p}
\title{Get -log10 p-value}
\usage{
get_neglog10_p(z = stop("Provide a z statistics"), n = stop("Provide a n"))
}
\arguments{
\item{z}{a z (or t) statistic}

\item{n}{the sample size used to estimate the beta and se}
}
\value{
Returns a -log 10 p-value
}
\description{
Get -log10 p-value from a provided z statistic and n
}
\examples{
get_neglog10_p(beta, se, n)

}
\author{
The internet (anon)
}
