% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_p}
\alias{get_p}
\title{Get p-value from Z-statistic}
\usage{
get_p(z)
}
\arguments{
\item{z}{a z (or t) statistic}
}
\value{
Returns a p-value (numeric)
}
\description{
Return p-value for a z (or t) statistic
}
\examples{
z = 10
get_p(z)

}
\author{
The internet (anon)
}
