% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_p_extreme}
\alias{get_p_extreme}
\title{Get extreme p-value}
\usage{
get_p_extreme(z)
}
\arguments{
\item{z}{a z (or t) statistic}
}
\value{
Returns a string (p-value)
}
\description{
Return p-value for any z (or t) statistic as a string
}
\examples{
z = 50
get_p_extreme(z)

}
\author{
The internet (anon)
}
