% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_p_neglog10}
\alias{get_p_neglog10}
\title{Get -log10 p-value (using Z only)}
\usage{
get_p_neglog10(z)
}
\arguments{
\item{z}{a z (or t) statistic}
}
\value{
Returns a -log 10 p-value
}
\description{
Get -log10 p-value from a provided z statistic
}
\examples{
z = 50
get_p_neglog10(z)
#>  [1] 544.0977

}
\author{
The internet (anon)
}
