% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_p_neglog10_n}
\alias{get_p_neglog10_n}
\title{Get -log10 p-value (using Z & N)}
\usage{
get_p_neglog10_n(z, n)
}
\arguments{
\item{z}{a z (or t) statistic}

\item{n}{the sample size used to estimate the z (or t) statistic}
}
\value{
Returns a -log 10 p-value
}
\description{
Get -log10 p-value from a provided z statistic and n
}
\examples{
z = 50
n = 100000
get_p_neglog10_n(z, n)
#>  [1] 537.9851

}
\author{
The internet (anon)
}
