% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_se}
\alias{get_se}
\title{Get Standard Error from Confidence Intervals}
\usage{
get_se(lci, uci, log = FALSE, denominator = 3.92)
}
\arguments{
\item{lci}{the lower confidence interval}

\item{uci}{the lower confidence interval}

\item{log}{log the CI values? (default=FALSE)}

\item{denominator}{the standard error of the sample mean (default=3.92 i.e., 1.96*2)}
}
\value{
Returns SE (numeric)
}
\description{
Return SE from CIs
}
\examples{
lci = 0.1
uci = 0.3
get_se(lci, uci)
#>  [1] 0.05102041

}
\author{
The internet (anon)
}
