% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_z}
\alias{get_z}
\title{Get Z-statistic from p-value}
\usage{
get_z(p, two_sided = TRUE)
}
\arguments{
\item{p}{a p-value}

\item{two_sided}{Logical, default=TRUE. Return two-sided z-value?}
}
\value{
Returns a Z-statistic (numeric)
}
\description{
Return a z-statistic from a given p-value
}
\examples{
p = 1e-10
get_z(p)

}
\author{
The internet (anon)
}
