% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_norm.R
\name{inv_norm}
\alias{inv_norm}
\title{Inverse normalize a variable}
\usage{
inv_norm(x = stop("Provide a vector"))
}
\arguments{
\item{x}{a numeric vector to be inverse normalized}
}
\value{
Returns a vector
}
\description{
Inverse normalize a variable (force a normal distribution)
}
\examples{
inv_norm(x)

}
\author{
Luke Pilling
}
