% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_themes.R
\name{theme_minimal_modified}
\alias{theme_minimal_modified}
\title{Modified theme_minimal for ggplots}
\usage{
theme_minimal_modified()
}
\value{
Returns a theme
}
\description{
From Albert Rapp's blog - nice adjustments to default theme minimal

https://github.com/AlbertRapp/3mw_newsletters/blob/master/37_Jul_19_2023/37_Jul_19_2023.qmd?utm_source=alberts-newsletter.beehiiv.com&utm_medium=newsletter&utm_campaign=3mw-create-your-own-ggplot-theme
}
\examples{
penguins = palmerpenguins::penguins |> 
  filter(!is.na(sex))

basic_plot = penguins |> 
  ggplot(aes(x = bill_length_mm, y = body_mass_g, fill = species)) +
  geom_point(shape = 21, size = 5, alpha = 0.85, color = 'grey10') +
  labs(
    x = 'Bill length (in mm)',
    y = element_blank(),
    fill = 'Species',
    title = 'Penguins from the Palmer Archipelago',
    subtitle = 'Penguin weights (in g)',
    caption = 'Data: {palmerpenguins} R package'
  )

basic_plot +
  theme_minimal(base_size = 16, base_family = 'Source Sans Pro') +
  scale_fill_manual(values = thematic::okabe_ito(6)) +
  theme_minimal_modified()

}
\author{
Albert Rapp
}
