% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assoc.R
\name{yv}
\alias{yv}
\title{Get outcome variables for using with `purrr::map2()`}
\usage{
yv(x_vars, y_vars)
}
\arguments{
\item{x_vars}{A vector of strings. The exposure variable names}

\item{y_vars}{A vector of strings. The outcome variable names}
}
\value{
Returns a vector of strings
}
\description{
The {purrr} package provides the wonderful `map()` and `map2()` functions. To get `map2()` to do each comparison need to provide a "long" version of the variable list
}
\examples{
x_vars = c("bmi","sbp")
y_vars = c("chd","t2d")
x_vars2 = xv(x_vars, y_vars)
y_vars2 = yv(x_vars, y_vars)

}
\author{
Luke Pilling
}
