% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{z_trans}
\alias{z_trans}
\title{Z-transform a variable}
\usage{
z_trans(x)
}
\arguments{
\item{x}{a numeric vector to be z-transformed}
}
\value{
Returns a vector
}
\description{
Z-transform a variable. Mean=0, SD=1. Maintains original distribution.
}
\examples{
z_trans(x)

}
\author{
Luke Pilling
}
