% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assoc.R
\name{get_assoc}
\alias{get_assoc}
\title{Get tidy model output of exposure(s) on outcome(s)}
\usage{
get_assoc(
  x,
  y,
  z,
  d,
  subset_d = TRUE,
  progress = TRUE,
  model = "lm",
  af = FALSE,
  note = "",
  scale_x = FALSE,
  scale_y = FALSE,
  inv_norm_x = FALSE,
  inv_norm_y = FALSE,
  extreme_ps = FALSE,
  ...
)
}
\arguments{
\item{x}{A string or vector of strings. The exposure variable name(s), found in `d`}

\item{y}{A string. The outcome variable name, found in `d` -- if model is 'coxph' then paste the survival object here e.g., 'Surv(time, event)' where `time` and `event` are variables in `d`}

\item{z}{A string. The covariate formula (e.g., " + age + sex"), found in `d`}

\item{d}{A data.frame or tibble. The data}

\item{subset_d}{Logical. Default is TRUE. Subset `d` before passing to analysis (much quicker when multiple exposures/outcomes used)}

\item{progress}{Logical. Default is TRUE. Show progress bar from {purrr} `map()` function (useful when multiple exposures/outcomes provided)}

\item{model}{A string. Default is "lm". The type of model to perform. Can also be "logistic" or "coxph"}

\item{af}{Logical. Default is FALSE. Is `x` categorical? I.e., include in formula as `as.factor(x)`}

\item{note}{A string. If you want to include a note like "All", "Males", "C282Y homozygotes" to describe the model or sample.}

\item{scale_x}{Logical. Default is FALSE. Apply scale() function to exposure?}

\item{scale_y}{Logical. Default is FALSE. Apply scale() function to outcome?}

\item{inv_norm_x}{Logical. Default is FALSE. Apply inv_norm() function to exposure?}

\item{inv_norm_y}{Logical. Default is FALSE. Apply inv_norm() function to outcome?}

\item{extreme_ps}{Logical. Default is FALSE. If p==0 then return "extreme p-values" as strings.}

\item{...}{Other `tidy_ci()` options}
}
\value{
Returns a tibble - summary statistics from a model
}
\description{
To easily get tidy model output (from linear or logistic GLM, or CoxPH) for a categorical or continuous exposure, including sample size (and N cases if logistic), outcome, and model info. Idea is to make quick loops easy.

For all exposures, it gets the N. For categorical exposures, the N is split by group, and a row is included for the reference category

Can provide multiple exposures and/or outcomes
}
\examples{
# for one outcome, equivalent to `tidy_ci(glm(weight ~ height +age+sex, d=ukb))` - with added `n`
get_assoc(y="weight", x="height", z="+age+sex", d=ukb)

# categorical exposure, binary outcome, and stratified analysis (with note)
get_assoc(y="chd", x="smoking_status", z="+age", d=ukb |> filter(sex==1), logistic=TRUE, af=TRUE, note="Males only")

# multiple exposures and/or outcomes
x_vars = c("bmi","ldl","sbp_0_avg")
y_vars = c("chd","t2d")
get_assoc(x=x_vars, y=y_vars, z="+age+sex", d=ukb, logistic=TRUE)

}
\author{
Luke Pilling
}
