% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_p_neglog10}
\alias{get_p_neglog10}
\title{Get -log10 p-value (using Z only)}
\usage{
get_p_neglog10(z, is_p = FALSE)
}
\arguments{
\item{z}{Numeric. A z statistic or p-value.}

\item{is_p}{Logical. Default=FALSE. Is this a p-value?}
}
\value{
Returns a -log 10 p-value
}
\description{
Get -log10 p-value from a provided z statistic (or p-value)
}
\examples{
z = 50
get_p_neglog10(z)

}
\author{
The internet (anon)
}
