% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{get_se}
\alias{get_se}
\title{Get Standard Error from Confidence Intervals}
\usage{
get_se(lci, uci, log = FALSE, denominator = 3.919928)
}
\arguments{
\item{lci}{The lower confidence interval}

\item{uci}{The lower confidence interval}

\item{log}{Log the CI values? (default=FALSE)}

\item{denominator}{The standard error of the sample mean (default=3.92... well actually, 3.919928 from `2*get_z(0.05)`)}
}
\value{
Returns SE (numeric)
}
\description{
Return SE from CIs
}
\examples{
lci = 0.1
uci = 0.3
get_se(lci, uci)

}
\author{
The internet (anon)
}
