% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{inv_norm}
\alias{inv_norm}
\title{Inverse normalize a variable}
\usage{
inv_norm(x)
}
\arguments{
\item{x}{a numeric vector to be inverse normalized}
}
\value{
Returns a vector
}
\description{
Inverse normalize a variable (force a normal distribution)
}
\examples{
summary(example_data$sbp)
sbp_in = inv_norm(example_data$sbp)
summary(sbp_in)

}
\author{
Luke Pilling
}
